Chore: Move expressjs fixture
