+++ pyproject.toml
@@ -45,6 +45,7 @@ pythonpath = [
   "examples/fastapi/charm/lib",
   "examples/go/charm/lib",
   "examples/expressjs/charm/lib",
+  "examples/flask-minimal/charm/lib",
   "lib",
   "src"
 ]
@@ -61,6 +62,7 @@ extend-exclude = '''
   |   examples/fastapi/charm/lib/.*
   |   examples/go/charm/lib/.*
   |   examples/expressjs/charm/lib/.*
+  |   examples/flask-minimal/charm/lib/.*
 )
 '''
