+++ tox.ini
@@ -71,6 +71,7 @@ commands =
       --skip {toxinidir}/examples/flask/lib \
       --skip {toxinidir}/examples/expressjs/charm/lib \
       --skip {toxinidir}/examples/expressjs/app/node_modules \
+      --skip {toxinidir}/examples/flask-minimal/charm/lib \
       --skip {toxinidir}/tests/integration/.mypy_cache
     # pflake8 wrapper supports config from pyproject.toml
     pflake8 {[vars]src_path} {[vars]legacy_src_path}
